refactor(sdk): update readme file & move styles under 'controls' directory
