+++ src/controls/TilesSelectorControl/styles/mapstyles_base.json
