+++ src/controls/TilesSelectorControl/styles/mapstyles_satellite.json
