+++ src/controls/TilesSelectorControl/styles/mapstyles_terrain.json
