+++ src/controls/TilesSelectorControl/styles/mapstyles_vectorial.json
