+++ src/controls/index.ts
@@ -6,6 +6,7 @@ export {
   default as IndoorControl,
   default as IndoorControlOptions,
 } from "./IndoorControl/IndoorControl";
+export { MapTiles } from "./TilesSelectorControl/styles/MapTiles";
 export { default as TilesSelectorControl } from "./TilesSelectorControl/TilesSelectorControl";
 export { default as CellPrecision } from "./GridControl/models/CellPrecision";
 export { default as DraftShapesControl } from "./DraftShapesControl/DraftShapesControl";
